binance_usdm_bot/requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
